+++ src/features/Guild/components/Overview/index.js
@@ -1,15 +1,20 @@
 // @flow
 
+import T from 'i18n-react';
+
 import Container from 'common/components/Container';
 import Redacted from 'common/components/Redacted';
 import SummaryProgress from 'common/layouts/SummaryProgress';
 import Card from 'common/layouts/CardWithTitle';
+import Grid from 'common/layouts/Grid';
+
 import styles from './styles.less';
 
 import type { Guild } from 'flowTypes';
+import { defaultGuild } from 'flowTypes';
 
 type Props = {
-  data: Guild,
+  data?: Guild,
 };
 
 const motdPlaceholder = 'Today we\'re going to go for an adventure...';
@@ -20,19 +25,19 @@ function parseNewLines (string = '') {
     : motdPlaceholder;
 }
 
-const Overview = ({ data = {} }: Props) => (
+const Overview = ({ data = defaultGuild }: Props) => (
   <Container>
-    <Card size="large" title="Message of the day" className={styles.motd}>
+    <Card size="large" title={T.translate('guilds.motd')} className={styles.motd}>
       <Redacted redact={!data.motd}>{parseNewLines(data.motd)}</Redacted>
     </Card>
 
-    <div className={styles.summaryContainer}>
+    <Grid>
       <SummaryProgress title="Level" max={69} current={data.level} />
       <SummaryProgress title="Favor" max={6000} current={data.favor} />
       <SummaryProgress title="Aetherium" max={25000} current={data.aetherium} />
       <SummaryProgress title="Resonance" subTitle={data.resonance} />
       <SummaryProgress title="Influence" subTitle={data.influence} />
-    </div>
+    </Grid>
   </Container>
 );